requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
outcome==1.3.0.post0
pandas==2.3.3
psycopg2-binary==2.9.11
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
